+++ HISTORY.md
@@ -1,5 +1,9 @@
 # History
 
+## UNRELEASED
+
+* Added an example for Typer.
+
 ## 0.3.2 (2021-08-24)
 
 * Added MatchBySubclass object matcher.