+++ HISTORY.md
@@ -4,6 +4,7 @@
 
 * (backward-incompatible) Renamed a matcher: `MatchByCallableAttribute` to `MatchByMethod`
 * Updated documentation for module and object rules.
+* Updated documentation for matchers.
 * Added an example for Typer.
 
 ## 0.3.2 (2021-08-24)