+++ HISTORY.md
@@ -6,6 +6,7 @@
 * Updated documentation for module and object rules.
 * Updated documentation for matchers.
 * Added an example for Typer.
+* Reorganized API documentation.
 
 ## 0.3.2 (2021-08-24)